Fix crashes when only sample type outputs are defined
- Crashes were due to missing trigger signal that was not read to
  self.iofile_eventdict, since trigger signal was not read unless
  explicitly defined to be an event type output

- On the other hand, explicitly defining triggers as event outputs
 caueses columns to overflow in the print file (i.e. num_cols is
 calculated incorrectly) in some cases

- The most convienient solution is to just read the output file once
BEFORE reading the individual IOs
