+++ spice/spice_iofile.py
@@ -298,179 +298,12 @@ class spice_iofile(iofile):
         else:
             pass
 
-    def parse_io_from_file(self,filepath,start,stop,dtype,labels,queue):
-        """ Parse specific lines from a spectre print file.
-
-        This is wrapped to a function to allow parallelism.
-        
-        [TODO] Convert to use kwargs
-        """
-        stack = [(label, None) for label in labels]
-        try:
-            nrows = stop - start
-            if nrows<0:
-                self.print_log(type='W', msg='Stop index smaller than start index in parse_io_from_file!')
-                nrows=None
-            arr=pd.read_csv(filepath,skiprows=start-1, nrows=nrows,
-                    sep='\s+', encoding='utf-8',engine='c',
-                    dtype='float',chunksize=1e6)
-            arr=pd.concat(arr).to_numpy()
-        except ValueError:
-            # This may happen if the print file
-            # does not round to zero, and the
-            # scientific format exponent may become
-            # over 100, where the space separators
-            # move on top of eachother. This
-            # adds the missing spacebars, which
-            # may fix the crash
-            try:
-                cmd=f'sed -i -E "s/([0-9])+([eE][+-]?[0-9]+)?[+-]/\\1\\2 /g" {filepath}'
-                self.print_log(type='I', msg="Running external command %s" %(cmd) )
-                subprocess.check_output(cmd,shell=True)
-                nrows = stop - start
-                if nrows<0:
-                    self.print_log(type='W', msg='Stop index smaller than start index in parse_io_from_file!')
-                    nrows=None
-                arr=pd.read_csv(filepath,skiprows=start-1, nrows=nrows,
-                        sep='\s+', encoding='utf-8',engine='c',
-                        dtype='float',chunksize=1e6)
-                arr=pd.concat(arr).to_numpy()
-            except: 
-                self.print_log(type='E',msg=traceback.format_exc())
-                self.print_log(type='F',msg='Failed while reading files for %s.' % self.name)
-        except: 
-            self.print_log(type='E',msg=traceback.format_exc())
-            self.print_log(type='F',msg='Failed while reading files for %s.' % self.name)
-        try:
-            n = 0
-            for i, label in enumerate(labels):
-                self.print_log(type='D',msg='Reading event output %s' % label)
-                if dtype=='complex': # Complex data has separate columns in file for real and imag parts
-                    try:
-                        temp=np.vstack((arr[:,0], arr[:,n+1]+1j*arr[:,n+2])).T
-                        n += 2
-                    except IndexError: # If the data isn't complex (might be the case if there is some real valued extract), parse as usual
-                        self.print_log(type='W', msg='Index overrange when reading data for output %s. Inferred datatype incorrect?' % label)
-                        temp = np.vstack((arr[:,0], arr[:,n+1])).T
-                        n += 1
-                else:
-                    temp=np.vstack((arr[:,0], arr[:,n+1])).T
-                    n += 1
-                stack[i] = (label, temp)
-            if queue!=None:
-                queue.put(stack)
-            else:
-                return stack
-        except:
-            self.print_log(type='E',msg=traceback.format_exc())
-            self.print_log(type='E',msg='Failed reading event output %s' % label)
-            if queue!=None:
-                queue.put(stack)
-            else:
-                return stack
-
     # Overloaded read from thesdk.iofile
     def read(self,**kwargs):
         """
         Function to read files associated with this spice_iofile.
         """
-        if self.iotype=='event':
-            file=self.file[0] # File is the same for all event type outputs
-            label_match=re.compile(r'\(([^)]+)\)') # Match one or more characters that are not ) and capture.
-            if self.parent.model in ['spectre','ngspice']:
-                os.system('sync %s' % self.parent.spicesimpath)
-                block_count=subprocess.check_output('grep -n \"time\|freq\" %s | sed \'s/^\([0-9]\+\):/\\1|/\'' % file, shell=True).decode('utf-8')
-                if not block_count: 
-                    # We couldn't find the block count, exit
-                    if os.path.isfile(file):
-                        self.print_log(type='F', msg='Missing header row(s) from .print file!')
-                    else:
-                        self.print_log(type='F', msg='.print file at %s doesn\'t exist!' % file)
-                blocks=block_count.split('\n') 
-                linenumbers=[]
-                labels=[]
-                # Parse linenumbers of header blocks
-                for block in blocks:
-                    parts=block.split('|')
-                    if len(parts) > 1: # Line should now contain linenumber in first element, ioname in second
-                        line = 0
-                        try:
-                            line=int(parts[0])
-                            linenumbers.append(line)
-                        except ValueError:
-                            self.print_log(type='W', msg='Couldn\'t decode linenumber from file %s' %  file)
-                        labelgrp=label_match.findall(parts[1]) # Parse IO labels (nodenames)
-                        if labelgrp:
-                            tmp = list(dict.fromkeys(labelgrp))
-                            labels.append(tmp)
-                        else:
-                            self.print_log(type='W', msg='Couldn\'t find IO on line %d from file %s' %  (line,file))
-
-                if len(labels) == len(linenumbers):
-                    try:
-                        numlines = int(subprocess.check_output("wc -l %s | awk '{print $1}'" % file,shell=True).decode('utf-8'))
-                    except FileNotFoundError as e:
-                        self.print_log(type='F', msg='Print-file doesn\'t exist! Invalid node names in saves statement?')
-                    except ValueError as e:
-                        self.print_log(type='F', msg='Print-file doesn\'t exist! Invalid node names in saves statement?')
-                    # Maximum number of concurrent open files. This may or may not help with "too many open files" -error.
-                    num_parallel = 50
-                    num_loops = int(np.ceil(len(linenumbers)/num_parallel))
-                    for it in range(num_loops):
-                        lnrange = range(num_parallel*it,min([num_parallel*(it+1),len(linenumbers)]))
-                        procs = []
-                        queues = []
-                        for k in lnrange:
-                            start=linenumbers[k] # Indexing of line numbers starts from one
-                            if k == len(linenumbers)-1:
-                                stop=numlines-1
-                            else:
-                                stop=linenumbers[k+1]-6 # Previous data column ends 5 rows before start of next one
-                            dtype=self.datatype if self.datatype=='complex' else 'float' # Default is int for thesdk_spicefile, let's infer from data
-                            nrows=stop-start
-                            if nrows<20e6:
-                                self.print_log(type='I',msg=f'Number of lines: {nrows}, reading with multiprocessing')
-                                queue = multiprocessing.Queue()
-                                queues.append(queue)
-                                proc = multiprocessing.Process(target=self.parse_io_from_file,args=(file,start,stop,dtype,labels[k],queue))
-                                procs.append(proc)
-                                proc.start() 
-                            else:
-                                self.print_log(type='I',msg=f'Number of lines: {nrows}, reading without multiprocessing')
-                                queue=None
-                                ret = self.parse_io_from_file(file,start,stop,dtype,labels[k],queue)
-                                for item in ret:
-                                    self.parent.iofile_eventdict[item[0].upper()]=item[1]
-                                self.print_log(type='I',msg=f'IO reading complete')
-                        for i,p in enumerate(procs):
-                            try:
-                                ret = queues[i].get()
-                                for item in ret:
-                                    self.parent.iofile_eventdict[item[0].upper()]=item[1]
-                                p.join()
-                            except KeyError:
-                                self.print_log(type='W', msg='Failed reading %s' % (ret[0]))
-                else:
-                    self.print_log(type='W', msg='Couldn\'t read IOs from file %s. Missing ioname?' % file)
-            elif self.parent.model == 'eldo':
-                # Parse signal headers
-                with open(file,'r') as f:
-                    for line in f.readlines():
-                        if line.startswith('# TIME') or line.startswith('# FREQ'):
-                            header = line.replace('# ','').replace('\n','').split(' ')
-                            break
-                arr = np.genfromtxt(file)
-                if len(header) != len(arr[0,:]):
-                    self.print_log(type='E', msg='Signal name and array column mismatch while reading event outputs.')
-                for col_idx,sname in enumerate(header[1:]):
-                    label=label_match.search(sname)
-                    if label:
-                        label = label.group(1)
-                        # Add to the event dictionary
-                        self.parent.iofile_eventdict[label.upper()]=np.hstack((arr[:,0].reshape(-1,1),arr[:,col_idx+1].reshape(-1,1))).reshape(-1,2)
-                    else:
-                        self.print_log(type='W', msg='Label format mismatch with \'%s\'.' %  (label))
-        elif self.iotype=='psfascii_pss' or self.iotype=='psfascii_pac':
+        if self.iotype=='psfascii_pss' or self.iotype=='psfascii_pac':
             self.psfasciiflag=True
             if not self.parent.model=='spectre':
                 self.print_log(type='F', msg='Only spectre supported for psfascii outputs')
@@ -549,7 +382,7 @@ class spice_iofile(iofile):
                         if trig.upper() not in self.parent.iofile_eventdict:
                             self.print_log(type='E',msg='Event data not found for trigger signal %s' % trig)
                         else:
-                            trig_event = self.parent.iofile_eventdict[trig]
+                            trig_event = self.parent.iofile_eventdict[trig.upper()]
                             tsamp = self.interp_crossings(trig_event,self.vth,256,self.edgetype)
 
                         # Processing each bit in the bus
